+++ src/ast.h
@@ -15,7 +15,7 @@
 class Node {
 public:
   virtual ~Node() = default;
-  virtual void write(std::ostream &os, int indent = 0) const = 0;
+  virtual void write(std::ostream &os, unsigned long indent = 0) const = 0;
   virtual void accept(ASTVisitor &visitor) = 0;
 };
 
@@ -33,7 +33,7 @@ using StmtPtr = std::shared_ptr<Statement>;
 class StatementList : public Node {
 public:
   std::vector<std::shared_ptr<Statement>> statements;
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
   explicit StatementList(const std::vector<StmtPtr> stmts)
       : statements(std::move(stmts)) {}
@@ -44,7 +44,7 @@ class ExpressionStatement : public Statement {
 public:
   ExprPtr expr;
   explicit ExpressionStatement(ExprPtr expr) : expr(std::move(expr)) {}
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
 };
 
@@ -52,7 +52,7 @@ class Number : public Expression {
 public:
   std::string value;
   explicit Number(const std::string &val) : value(val) {}
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
   AlohaType::Type getType() const override { return AlohaType::Type::NUMBER; }
 };
@@ -61,7 +61,7 @@ class Boolean : public Expression {
 public:
   bool value;
   explicit Boolean(const bool &val) : value(val) {}
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
   AlohaType::Type getType() const override { return AlohaType::Type::BOOL; }
 };
@@ -70,7 +70,7 @@ class AlohaString : public Expression {
 public:
   std::string value;
   explicit AlohaString(const std::string &val) : value(std::move(val)) {}
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
   AlohaType::Type getType() const override { return AlohaType::Type::STRING; }
 };
@@ -84,7 +84,7 @@ public:
   UnaryExpression(const std::string &oper, ExprPtr expr)
       : op(oper), expr(std::move(expr)), type(AlohaType::Type::UNKNOWN) {}
 
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
   AlohaType::Type getType() const override { return expr->getType(); }
 };
@@ -100,7 +100,7 @@ public:
       : left(std::move(lhs)), op(oper), right(std::move(rhs)),
         type(AlohaType::Type::UNKNOWN) {}
 
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
   AlohaType::Type getType() const override { return left->getType(); }
 };
@@ -113,7 +113,7 @@ public:
       : name(name), type(t) {}
   explicit Identifier(const std::string &name)
       : name(name), type(AlohaType::Type::UNKNOWN) {}
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
   AlohaType::Type getType() const override { return type; }
 };
@@ -127,7 +127,7 @@ public:
   explicit Declaration(const std::string &varName,
                        std::optional<AlohaType::Type> type, ExprPtr expr)
       : variableName(varName), type(type), expression(std::move(expr)) {}
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
 };
 
@@ -138,7 +138,7 @@ public:
   AlohaType::Type type;
   FunctionCall(std::shared_ptr<Identifier> &funcName, std::vector<ExprPtr> args)
       : funcName(funcName), arguments(std::move(args)) {}
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
   AlohaType::Type getType() const override { return type; }
 };
@@ -147,7 +147,7 @@ class ReturnStatement : public Statement {
 public:
   ExprPtr expression;
   explicit ReturnStatement(ExprPtr expr) : expression(std::move(expr)) {}
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
 };
 
@@ -160,7 +160,7 @@ public:
               std::shared_ptr<StatementList> elseBr)
       : condition(std::move(cond)), thenBranch(std::move(thenBr)),
         elseBranch(std::move(elseBr)) {}
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
   bool has_else_branch() const { return elseBranch != nullptr; }
 };
@@ -171,7 +171,7 @@ public:
   std::shared_ptr<StatementList> body;
   WhileLoop(ExprPtr cond, std::shared_ptr<StatementList> b)
       : condition(std::move(cond)), body(std::move(b)) {}
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
 };
 
@@ -185,7 +185,7 @@ public:
           std::shared_ptr<Declaration> inc, std::vector<StmtPtr> b)
       : initializer(std::move(init)), condition(std::move(cond)),
         increment(std::move(inc)), body(std::move(b)) {}
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
 };
 
@@ -210,7 +210,7 @@ public:
            std::shared_ptr<StatementList> b)
       : name(funcName), parameters(std::move(params)), returnType(retType),
         body(std::move(b)) {}
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
 };
 
@@ -218,7 +218,7 @@ public:
 class Program : public Node {
 public:
   std::vector<std::shared_ptr<Node>> nodes;
-  void write(std::ostream &os, int indent = 0) const override;
+  void write(std::ostream &os, unsigned long indent = 0) const override;
   void accept(ASTVisitor &visitor) override { visitor.visit(this); }
 };
